Add files via upload
I'm not entirely sure if my commits have been saving, so this is an alternate method to see if it has been working or not. If not, i'll have to figure something out as to how I can show the development of my code. I still have the comments I made as to what I changed and what I planned to do next time, however there isn't technically code proof. If it turns out to be true, I'll ask what there is I can do to show the changes, and if the comments are enough to prove it.