Add files via upload
Started Working on my button code for the title screen. I'm planning on having a title screen as to add a couple more elements into the game, as well as it being a nice general feature to any game, a feature that is heavily appreciated by users. I haven't gotten anything to work yet, but as of right now, I'm planning on putting the starter screen in a class, which will run until the player selects the button. Upon the player pressing the button, the start screen class will end and the game class will run. I don't actually know if this will work yet, but I'll see once its done.